Embed tokenizer interface in strtok3 (used to be token-types). Rename lib/type to lib/types.
