+++ package.json
@@ -46,14 +46,14 @@
     "remark-cli": "^7.0.1",
     "remark-preset-lint-recommended": "^3.0.3",
     "source-map-support": "^0.5.16",
+    "token-types": "^1.2.2",
     "ts-node": "^8.5.2",
     "tslint": "^5.20.1",
     "typescript": "^3.7.4"
   },
   "dependencies": {
     "debug": "^4.1.1",
-    "then-read-stream": "^2.0.8",
-    "token-types": "^1.2.2"
+    "then-read-stream": "^2.0.8"
   },
   "keywords": [
     "tokenizer",
